hostがリクエストを受け取り、処理するように
